requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
gunicorn==22.0.0
SQLAlchemy==2.0.30